Refactor history repo namespace to HistoryRepo
Updated namespace from IHistoryRepository to HistoryRepo in all related files, including WeatherService.cs, HistoryRepository.cs, and IHistoryRepository.cs. Adjusted using statements and namespace declarations to match the new structure.
